+++ smarttubetv/src/main/java/com/liskovsoft/smartyoutubetv2/tv/ui/widgets/embedplayer/EmbedPlayerView.java
@@ -281,8 +281,18 @@ public class EmbedPlayerView extends PlayerView implements PlaybackView {
     }
 
     @Override
+    public void openHlsUrl(String hlsPlaylistUrl, MediaItemFormatInfo formatInfo) {
+        mExoPlayerController.openHlsUrl(hlsPlaylistUrl, formatInfo);
+    }
+
+    @Override
     public void openUrlList(List<String> urlList) {
         mExoPlayerController.openUrlList(urlList);
+    }
+
+    @Override
+    public void openUrlList(List<String> urlList, MediaItemFormatInfo formatInfo) {
+        mExoPlayerController.openUrlList(urlList, formatInfo);
     }
 
     @Override